Core 8860 Refactor App integration Window to use the new `Manage Tools` UI

+++ de-lib/src/main/java/org/iplantc/de/client/models/tool/Tool.java
@@ -59,5 +59,5 @@ public interface Tool extends HasId, HasDescription, HasName {
     boolean isPublic();
 
     @PropertyName("time_limit_seconds")
-    long getTimeLimit();
+    Integer getTimeLimit();
 }
